Fix preview error. The error cause is initialization of systemUiController.
