+++ ui/common/src/main/java/dev/mkeeda/spaceship/ui/common/theme/Theme.kt
@@ -5,9 +5,6 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.SideEffect
-import androidx.compose.ui.graphics.Color
-import com.google.accompanist.systemuicontroller.rememberSystemUiController
 
 private val DarkColorPalette = darkColors(
     primary = Gray400,
@@ -30,15 +27,6 @@ fun SpaceshipTheme(
         DarkColorPalette
     } else {
         LightColorPalette
-    }
-
-    val systemUiController = rememberSystemUiController()
-    val useDarkIcons = colors.isLight
-    SideEffect {
-        systemUiController.setSystemBarsColor(
-            color = Color.Transparent,
-            darkIcons = useDarkIcons
-        )
     }
 
     MaterialTheme(
